+++ changelog.md
@@ -5,6 +5,13 @@ Version History
 -----------
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
+### What's new in version 1807.1?
+
+* Supports App Service 1.3 (Update 3) - required SQL PS to enable 'Contained DB Authentication'
+* For ASDKs that are registered, the script will now download all Microsoft VM Extensions
+
+![Existing files](deployment/offline/media/VMextensions.png)
+
 ### What's new in version 1807?
 
 * Support for AzureStack PowerShell Module 1.4.0